Set process_options in different file
