+++ main.rb
@@ -1,4 +1,6 @@
 require_relative 'app'
+require_relative 'process_options'
+require_relative 'display_menu'
 
 def display_menu
   puts 'Please choose an option by entering a number'
@@ -10,30 +12,6 @@ def display_menu
   puts '6 - List all rentals for a given person id'
   puts '7 - Exit'
 end
-
-# rubocop:disable Metrics/CyclomaticComplexity
-def process_option(option, app)
-  case option
-  when 1
-    app.list_books
-  when 2
-    app.list_people
-  when 3
-    app.create_person
-  when 4
-    app.create_book
-  when 5
-    app.create_rental
-  when 6
-    app.list_rentals
-  when 7
-    puts 'Thank you for using this app'
-    exit
-  else
-    puts 'Invalid option. Please choose a number between 1 and 7.'
-  end
-end
-# rubocop:enable Metrics/CyclomaticComplexity
 
 def main
   app = App.new
